requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# OCR Engines
paddleocr>=3.3.0
paddlepaddle>=2.5.0
easyocr>=1.7.0
pytesseract>=0.3.12

# LLM Integration
openai>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Image Processing
scikit-image>=0.21.0
albumentations>=1.3.0

# Metrics
rapidfuzz>=3.0.0
